+++ CPoint.h
@@ -34,7 +34,7 @@ public:
     // Contrat :
     //
 
-	bool Move(int dX, int dY);
+	bool Move(long dX, long dY);
     // Mode d'emploi :
     //
     // Contrat :
